backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
anthropic>=0.40.0
httpx>=0.27.0
pyautogui==0.9.54
mss==9.0.2
Pillow==10.4.0
pydantic==2.8.0
python-dotenv==1.0.1
openai>=1.40.0
tavily-python>=0.5.0
playwright>=1.45.0
